fix hanging tests
This fixes hanging tests. It involves splitting some files into very
specific roles, such as `index.ts` for our entry point, `app.ts` for our
main route handling, and `routes.ts` for our prefixed
routes (`/api/v1`). As a result of this split, `listen` is not being
called during test startup. Instead, the server is started by Supertest.
Supertest should handle this just fine (their docs indicate it should),
but there is likely a relationship between controlled startup and
promise based testing. See
https://github.com/visionmedia/supertest/issues/697 for some more
information about this.
